+++ RELEASE_NOTES.md
@@ -1,1 +1,2 @@
-- Postpone command added
+- Update some dependencies 
+- Health controller added